Refactoring - separating v1/user.py for readability & design purpose
