+++ api/routes/v1/__init__.py
@@ -1,1 +1,4 @@
 from api.routes.v1.user import user_blueprint
+from api.routes.v1.index import homepage_blueprint
+from api.routes.v1.seed import seed_blueprint
+from api.routes.v1.message import message_blueprint
