+++ api/routes/v1/user.py
@@ -1,79 +1,10 @@
 from api.controllers.token import refresh
-from api.seed.seeder import seed
 from api.database.db import DataBase
 from api.controllers.user import *
 from api.validations import *
 
 user_blueprint = Blueprint("user", __name__)
 collection = DataBase.get_instance()
-
-
-@user_blueprint.route('/')
-def home_page():
-    """
-    Home page
-    :return: message response
-    """
-    return "Messaging System - REST API!"
-
-
-@user_blueprint.route('/seed', methods=['POST'])
-def seed_db():
-    """
-    Performs seeding to mongo db
-    :return: message response
-    """
-    return seed()
-
-
-@user_blueprint.route('/messages', methods=['GET'])
-@authorize_required
-def get_messages(user_data):
-    """
-    Get all read messages / unread messages
-    Postman exam: WEB_ROUTE/messages?only_unread=True
-    :param user_data: user details
-    :return: all messages for a specific user
-    """
-    return read_all_messages(collection, user_data)
-
-
-@user_blueprint.route('/messages/<string:messageId>', methods=['GET'])
-@authorize_required
-def get_single_message(user_data, **kwargs):
-    """
-    Get message by id
-    Postman exam: WEB_ROUTE/messages/MESSAGE_ID_FROM_MONGO_DB
-    :param user_data: user details
-    :return: Details of one message
-    """
-    return read_message(collection, kwargs["messageId"], user_data)
-
-
-@user_blueprint.route('/messages/<string:messageId>', methods=['DELETE'])
-@authorize_required
-def delete(user_data, **kwargs):
-    """
-    Delete specific message by id
-    Postman exam: WEB_ROUTE/messages/MESSAGE_ID_FROM_MONGO_DB
-    :param user_data: data details
-    :return: response / feedback
-    """
-    return delete_message(collection, kwargs["messageId"], user_data)
-
-
-@user_blueprint.route('/messages', methods=['POST'])
-@authorize_required
-@validate_request("message", "create")
-def set_message(user_data, message):
-    """
-    Post message
-    Postman exam: WEB_ROUTE/messages
-    :param user_data: user details
-    :param message: user message
-    :return: message id
-    """
-    return write_message(collection, user_data, message)
 
 
 @user_blueprint.route('/user', methods=['POST'])
